CHORE(migration): added database migration scripts
This commit adds important database migration scripts.  This includes a
shell script for importing data from a bhima-1.X dump file and schema
upgrades to be run on a live bhima-1.X database.

Case 1: To upgrade from a dumped file (such as the dropbox backup), the
correct approach is to unzip the file and run the script
./sh/migrate-previous-dump.sh on the file.  This script will"
  a) Change all CHARSETs to UTF-8
  b) Change all ENGINEs to InnoDB
  c) Reset all AUTO_INCREMENTs to 0

Case 2: To migrate the bhima-1.X schema to the bhima-2.X schema, use the
SQL files in update/schema-v2.sql   This SQL file uses temporary tables
to ensure that all data is properly copied over to a new schema.