TEST(cash): added /cash integration tests
This commit adds ~12 failing /cash API endpoint tests for the caution/invoice
payment functionality of the cash module.

Additionally, a test helper file (server/test/api/helper.js) is added to
handle routine test parameters without removing readability.  The /cash
API is the first to leverage this helpers and can be used as a reference
while the work is expanded to other files.