WIP(cash): moved cashbox selection out of /cash
This commit removes cashbox selection from the /cash controller and puts
it in its own controller/template.  This greatly reduces the complexity
in the cash controller, but requires a better solution for state.

The routes now look like this:
 - /cash matches to the cashbox selection page
 - /cash/:id attempts to load a cashbox from the server matching :id.
   If none exists, reroutes to the /cash page.

In the future, this state-like behavior should be encapsulated in
something like ui-router.