CHORE(cash): Finalize receipt rendering
This commit fixes the cash receipt rendering. The following changes have
been made:
  1. ExchangeRateService simplified to use the same code to lookup a
  previous exchange rate.  Huge improvement to maintainability and
  testability.
  2. English translations added.  French translations are a TODO.
  3. The PatientService has been brought up to 2.X standards.  Notably
    - uses util.unwrapHttpResponse()
    - instantiated by angular.service
  4. Cleaner layout for receipt modal HTML
  5. Loading indicator while data loads.

Feature Request: Currencies.round(currencyId, amount) to round
currencies to their nearest min monentary unit.  Since we store values
in a single currency, we need a way to reliably represent real-world
payments when converting from stored values.  Currently, it is psosible
for the cash receipt to read "102.45FC", which obviously is not correct.