CHORE(lib): deprecate util, validate, sanitize
This commit deprecates and deletes the vast majority of code in the
following files:
 - lib/util.js
 - lib/validate.js
 - lib/sanitize.js

Any remaining methods have been wrapped in NodeJS's util.deprecate()
method to indicate that they should not be used.  These methods will be
completely removed in the coming month.

Additionally, lodash has been added as a dependency.  This serves to
fill some of the needs left by the deletion and deprecation of our lib
functions.

Closes #32.