test(/locations): integration tests for /locations
This commit rewrites and adds tests for the /locations API.  The
location.js file has been renamed, and many of the route methods have
been named according to new naming conventions.

This commit also adds a new test helpers "listed(res, number)" to assert
that the API successfully returned an array of length {number}.  It
should be useful for future APIs going forward.