(FIX) Resolve minor bug with tree units cache
Tree directive would attempt to set the open state on a unit even if the
unit was no longer a valid route according to the current user. As this
unit no longer existed a console error would be reported.

Update the directive to first check if the cache is a valid unit and if
not remove it from the record.