chore(e2e): polish wrappers for bh-location-select
This commit refines the wrappers for the bh-location-select.  Multiple
location selects are now able to be located on the same page, namespaced
by a user-provided id.  The get() and set() methods have been
implemented to take in and return an array of UUIDs, respectively.

The patient registration e2e tests now include an example of this
component being used.