feat(reference): implement prototype reference modal
This commit implements a reference prototype modal that can be embedded
in any controller to retrieve the uuid of any previous transaction just
given the document reference.

Note: this only implements the UI.  The $http requests are mocked via
the $timeout inject until a backend is created for these queries.