chore(vouchers): implement voucher page w/ modal
This commit cleans up the old journal voucher UI, migrating it to
something that looks like bhima-2.X standards.  It is incomplete though.

I've added in a button to the JV page to open the ReferenceLookupModal.
This gives us a chance to experiment with UX while the backend framework is
under construction.