chore(build): build vendor into single file
This commit builds the vendor file into a single JS file.  This means
that our index.html file retrieves only two files:
 1) vendor.min.js
 2) bhima.min.js

IMPORTANT: This commit also removes the dependency on Slickgrid - it is
no longer shipped to the client at all.  For bhima-2.X, SlickGrid will
be replaced with UI Grid, so this should not cause any disruption in
workflow.

Closes #59.