feat(locations): implement add location modal
This commit implements the scaffolding for an add location modal.  The
goal is to be able to add any type of location easily and quickly to the
bhima database using a modal interface.  There is a link between the
location select component and the modal for easy opening/closing.

TODO - make the link optional in the component options.