Invoice items set and apply price list to lineitem
Price list fetched based on debtitor record - if a price list is present
it will be applied at the line item confirmation step. Proof of concept
information popover is used to demonstrate that a price list has been
applied to ensure the user knows that it is not a bug. This design
should be iterated upon.