feat(bhSubmit): submission hook which toggles loading state
This commit creates a custom directive for form submission to be used on
ngForm directives.  It binds a property `$loading` to the FormController
to signify when the form is in a loading state.  This property can be
used within the form to toggle UI views (such as loading indicators,
etc).

The `bh-submit` directive must receive a callback function that returns
a promise.  See the CashController.submit() method for an example.

USAGE:
```html
<form name="SomeCoolForm" bh-submit="SomeCtrl.submit()">

  <!-- form elements -->

  <!-- this <p> will only be shown when the form is loading -->
  <p ng-if="SomeCoolForm.$loading">
    I am loading!! :)
  </p>

</form>
```
```js
// inside SomeCtrl.submit()

$ctrl.submit = function submit() {
  // do some preprocessing steps
  return $http.post('/some/route', data)
  .catch(function (error) {
    $ctrl.HttpError = error;
  });
}
```

This feature is up for review by the team.  If it seems to be the best
way to handle form (loading) state, then we can proceed to use it for
loading indicators and buttons.