chore(languages): implement a LanguageService
This commit implements a LanguageService, responsible to managing the
languages and locales throughout the application. In order to do this,
several changes have been made:
 1. The database `language` table now contains the locales for
 angular-moment and tmhDynamicLocale.  This requires a database rebuild.
 2. Three controllers responsible for languages now use LanguageService.
 These are:
   1. ApplicationController
   2. LoginController
   3. SettingsController

USAGE
If you want to use the LanguageService to display or set languages from
a controller, the API is quite simple.  Languages are loaded from the
database on startup, and can be accessed via the `read()` function as
shown below.

```js
angular.module('bhima.controllers')
.controller('SomeController', [ 'LanguageService', function (Languages) {
  var vm = this;

  // reading languages from the service
  Languages.read()
  .then(function (languages) {
    vm.languages = languages;
  });

  // setting the application language based on a locale key
  Languages.set(key);

  // get the current language
  var languageKey = Langauges.get();
}]);
```

Closes #160.