chore(tree): make bhima-nav a component
This commit migrates the `bhimaNav` directive to a component
architecture.  The navigation component is now called `bhNavigation` and
has been migrated to the `components/` directory.

One additional improvement has been implemented from #44.  The tree now
searches through all paths for a match to the current url on change
using a `String.includes()` to search for matches.  Credit to @sfount
for this suggestion.