chore(db): schema FK fixes, voucher API fixes
This commit adds foreign keys into the `voucher` table in the database.
It brings the columns up to bhima-2.X standards and adds a `created_at`
timestamp.

This commit also cleans up the Voucher API significantly, and fixes the
tests appropriately.  The Voucher API now accepts a JSON object with a
single `voucher` property, with the list of voucher items under the
`voucher.items` sub-property.  This should make the client-side code
much cleaner.

Additional changes:
 1. The server will assign UUIDs if they do not exist on the voucher
 and/or voucher items.
 2. A utility method was written for mapping arrays of json objects to
 simple, ordered arrays.