chore(db): build db from two files
This commit remove synt.sql and ensures that the database is being built
only from the schema and test data files. This should make it much
easier and clearer when adding new data to the database.

All futher schema changes should occur in schema.sql, and all additional
data should be added to test.sql.

Closes #186.