feat(bhSubmitButton): loading-state aware btn
This commit adds the bhSubmitButton and a demonstration of it on the
cash page.  The button's signature is super simple:
```html
<bh-submit-button loading-state="ParentForm.$loading">
</bh-submit-button>
```

Ideally, developers should use it with a parent form that has the
`bhSubmit` directive attached to it, for the easy use.  However, it can
be used independently of the `bhSubmit` directive, provided the loading
state is clearly accessible to it.