Prototype initial core posting journal files
This commit introduces the initial files for the 2.x client side
posting journal. It proposes a structure for how the core UI grid
should be configured and sets up the testing framework.

Testing - A page object is used to abstract methods on the page
components (primarily the grid), currently this provides a method to
verify the total number of rows that can be accesed by the test runner
(journal.spec.js). This method is taken from the test utilities provided
by Angular UI Grid, this is a fully featured set of tools however the
actual expect behaviour is configured for a different testing framework.

Service Infrastructure - A service 'TransactionService' has been
created to interface with the server journal end point and provide any
utility methods required to manipulate transactions. Seperating models
and functionality into services will be a key aspect of how the
controller and module achieves modularity.

Routing - app.js is updated such that /#/journal points to the new 2.x
journal files. Once the updated posting journal is stable it will
completely replace the previous implementation; this should be reflected
in the file structure.