Propose modular service based architecture - Sort
This commit proposes a service based approach for each of the utilities the
journal will require. In this proposal sorting is implemented by creating the
JournalSortingService. This service is responsible for
- Configuring the global gridOptions object
- Providing helper functions for complex sorting

TransactionService was renamed to journal/TransactionService, this service
will still be available to any controllers that require however the current
implementation is directly intended for the posting journal.