Implement pagination service
This commit adds the files required for the journal pagination service,
providing a number of global configurations as well as helper methods to
settup correct pagination.

JournalPagination can be configured to specify if the service should use
external (custom) pagination (as defined in the
JournalPaginationService) or to use the default Angular UI Grid
pagination. Custom pagination (driven by the method `fetchPages`)
respects transactions and will not split a transaction across multiple
pages; this means the page size is dependent on the current page of
trnnsactions.

Currently isExternalPagination is set to false; until the jounral
requires features from the custom pagination this will be set - there
are many corner cases to consider with pagination (sorting, selection,
cacheing) that will require a very careful tested design.

- Updated column definitions go group (and sort) on transactions by
  default. This gives the core view into the posting journal.