chore(billing services): client structure changes
This commit changes the structure of the client to allow the abstract
state better control over child states, and access to the $state
service.  This will make it possible in the future to hide/show buttons
in the headercrumb based on the $state.current.

Importantly, this commit introduces the pattern of
 - /list
   - list.js
   - list.html
for default lists.  In the future, we might want to change this to
 - /index
   - index.js
   - index.html
if it makes more sense (such as most pages not using lists for their
interface).

Finally, the grid now has loading indicators and error indicators, which
should ease user experience in low-bandwidth environments.