(CHORE) Minor comment refacotr; grid button style
This commit refactors some of the comments in the users and permissions
controller to make them compatible with JSDoc.

Edit and Permissions buttons in the UI grid are now stored in their own
columns and are styled by bootstrap. Notible configuration changes
- Columns are given names with displayName set to ''
- Columns cannot be sorted
- None of the grid columns have menus

Inline templates should be extracted to seperate HTML template files in
a further refactor.