(CHORE) Update all required fields w/ ng-messages
This commit introduces ng-messages into the patient registration module.
Every form that is required or has validation (min/max) includes
ng-messages as a help block.

A file specifically for date validation was added included English
translations - this allows messages for min and max to be much more
relevent to the field ('This date is too far in the past' vs. 'This
value is too small').
This can be used by referencing
`partials/templates/dates.messages.tmpl.html`

Updated the location select component to use
templates/messages.tmpl.html and include the help block within the scope
of a has-error feedback group (ensuring it is coloured correctly)