(Chore) Remove $scope; packaging req in service
This commit cleans up the patient registration controller. The $scope object
is no longer required as an import and all references to it are
removed.

Data is now packaged for the server API end point in the
Patients service instead of the patient registration controller.

Removed references to old location selection utilities (which also used
the $scope variable).

Updated tests to no longer expect an alert but now check for the more
standard ng-messages.