(REFACTOR) Handle exceptions w/patient reg. module
This commit implements improved exception handling into the patient
registration controller.

If no debtor groups are downloaded this is reported at the debtor group
select (with a uib-alert); this must be explained as this is a required
field for the page to submit a new patient request.

Added a uib-alert at the top of the page that displays information if an
exception object is found on the view model. Updated relevent services
to pass the latest standard of translatable keys to ensure error
propegation is handled correctly at each stage.

$anchorScroll service is used to scroll to the error alert if the screen
is not currenlty focussed on this element; initial call is not included
in $apply block so for now the call is wrapped in a $timeout request.
All of this functionality should be wrapped in a service to reduce
number of requires for the controller.