(Feature) ScrollTo wrapper service; refact. names
This commit introduces a very simple wrapper for scrolling to elements
in angular. This is primarily to reduce the number of imports a
controller has to make to introduce the scrolling feature.
$location
$anchorScroll
$timeout (depending on where the scroll is invoked from)

The scroll method is wrapped in $timeout to ensure that it is called
from an $apply block - in a number of cases the scrolling method would
not be applied unless it was called a number of times.

Renamed the Patient and Debtors service to reflect the latest service
naming standards