feat: initial confirm modal implementation
This commit adds in a modal replacement for $window.confirm() for better
security.  See #150 for reference.

The modal is available in the ModalService, which also contains some
TODOs for later development.  The service exposes a single method,
`confirm()` that will resolve with a boolean value, depending on what
the user click - "true" for accept and "false" for cancel.