(FIX) Bug caught by failing supplier e2e test
This commit resolves a bug caught by a failing e2e that was allowed into
the master branch. In future processes should be improved such that a
pull reqest must pass all e2e to be allowed to merge.

The 'note' field on the supplier entity was capped at 50 characters.
This meant that the e2e providing a longer note was failing and causing
a 500 error.

The solution:
* Increase the size of the 'note' field for suppliers by making it TEXT
* Implement the 'DB_DATA_TOO_LONG' error with the server error, the
  error should be handled as a 400 (Bad Request)