(Feature) Handle corner cases and exceptions
This commit adds validation checking, as $q does not support cascading
.catch blocks a $q.deferred object is used to catch errors in the
directives scope.

If an error is thrown outside of the knowledge of the directive (no
internet, 500 server error) the directive will set invalid with a
provided key - this key is then translated and reported through an
additional record in ng-messages.