(Example) Implement bh-unique w/ hospital no.
This commit provides an example for the bh-unique directive. Patient
registration hospital numbers are required to be unique by adding the
attribute:
`bh-unique="/patients/hospital_number"`

The previous (hardcoded) version of async validation on hospital numbers
has been removed.

A proposal for the `exists` API powering the unique directive/ service
has been proposed. It is as follows:
`GET /entity/attribute/:id/exists`

More information and examples can be found in the services documenation.

Updated the /patients route with the proposed API pattern.