(Feature) Infrastructure for specifying renderer
This commit introduces the proposed render API. Any renderer can be
written to accept `data` and `options` parameters - the renderer must
return a promise that will resolve with the compiled report.

This example provides a JSON render intergace that simply returns the
data object given this standard interface.

Render target are attached to a `supportedRender` object in the report
controller and any key provided not attached to this object is sent back
to the client as a `400 Bad Request`.

Two additional integration tests added to account for additional end
point functionality:
* A valid render target should return a correctly formatted report
* An invalid render target should return `400 Bad Request`