chore(journal): updates the journal schema
This commit updates the journal schema to 2.x standards.  It brings
features like binary uuids and better naming conventions to the journal
backend.

The following changes are landing:
 1. removed `cost_ctrl_id`
 2. removed `session_id`
 3. removed `posting_session` table (for now)
 4. `uuid char(36)` --> `uuid binary(16)`
 5. `doc_num char(36)` --> `record_uuid binary(16)`
 6. `inv_po_id char(36)` --> `reference_uuid binary(16)`
 7. `deb_cred_uuid char(36)` --> `entity_uuid binary(16)`
 8. `deb_cred_type`  --> `entity_type`

Notably, this removes the `posting_session` table.  This is not a
permanent change, but the posting_session process should be designed by
the team before implementation.  For now, I've gutted it.

Closes #112.