chore: shuffle debtors/groups files on the server
This commit shuffles the debtors and debtorGroups files on the server,
changing their location to better capture the relationship between them
and allow other files to be added in next to them.

The new structure is
 - `controllers/debtors/index.js`
 - `controllers/debtors/groups/index.js`

Additionally server now references only the debtor groups module for
debtor groups routes (originally deprecated by @mbayopanda).  There is
still a lot of work remaining to get testing up and running again.