Propose debtor group card layout;+filter,sort,page
This commit introduces a number of utilites for the debtor groups
configuration page:
* Optional filtering on debtor group name
* Pagination for many debtor groups
* Semantic sorting on name and date created; listed as
  * Newest
  * Oldest
  * Name [A-Z]
  * Name [Z-A]
* Proposed card layout with mock information

The bootstrap .container class has been overriden favouring max-width
vs. width on different screen sizes. This is because BHIMAs tree
navigation takes up space that is not accounted for by the screen width
calculation (causing it to take up to much space with a specified
width). This may be replaced in an upcoming refactor.