Base printing styles + invoice mock; inject option
This commit introduces the first version of the printing styles that
will allow the selective printing of just modals (Ideal for receipts/
reports). These styles must be finalised and moved out of the in-line
receipt.

Options are now passed to the receipt modal controller, this is favoured
over many paremeters as there could be any number of configurations that
go into requesting a receipt report.
Currently the following options are respected by the receipt controller:
* `title` - this sets the first headercrumb element, it should be the
name of the receipt entity
* `identifier` - this is used to display the current ID of the if
receipt object in the headercrumb
* `template` - depending on the renderer used, a template is fetched to
bind to the data fetched from the server