chore: binary uuid migration (#289)
* chore(journal): updates the journal schema

This commit updates the journal schema to 2.x standards.  It brings
features like binary uuids and better naming conventions to the journal
backend.

The following changes are landing:
 1. removed `cost_ctrl_id`
 2. removed `session_id`
 3. removed `posting_session` table (for now)
 4. `uuid char(36)` --> `uuid binary(16)`
 5. `doc_num char(36)` --> `record_uuid binary(16)`
 6. `inv_po_id char(36)` --> `reference_uuid binary(16)`
 7. `deb_cred_uuid char(36)` --> `entity_uuid binary(16)`
 8. `deb_cred_type`  --> `entity_type`

Notably, this removes the `posting_session` table.  This is not a
permanent change, but the posting_session process should be designed by
the team before implementation.  For now, I've gutted it.

Closes #112.

* chore: shuffle debtors/groups files on the server

This commit shuffles the debtors and debtorGroups files on the server,
changing their location to better capture the relationship between them
and allow other files to be added in next to them.

The new structure is
 - `controllers/debtors/index.js`
 - `controllers/debtors/groups/index.js`

Additionally server now references only the debtor groups module for
debtor groups routes (originally deprecated by @mbayopanda).  There is
still a lot of work remaining to get testing up and running again.

* chore: fix debtors/groups tests

This commit fixes a few typo bugs in the debtor and debtor groups tests
which prevented successful execution and killed the server.  The
remaining test failures are logical failures (no ability to post or scan
the posting journal) and will be fixed in subsequent commits.

* chore(server): rm deprecated routes/controllers

This commit removes a few extra 1.x routes and controllers.  Some
of these are duplicates (subsidy-subsidies), others have simply been
superseded by better designs (conventions).

Closes #265.

* chore: migrate schema and data sql to binary(16)

This commit changes the entire database schema to use binary(16) UUIDs.
The database schema has been completely migrated and the test dataset
uses HUID() to convert uuids to binary.
Closes #262.

* chore: server api migration for BIN(16) uuids

This commit migrates the server apis to using binary uuids.  Every
server-side controller that interacts with tables that have uuids in
them has been updated to convert between binary and hexidecimal as
needed.

* chore: ensure end-to-end tests pass

This commit ensures that all end to end tests pass.  In order to meet
this objective, I had to introduce `server/finance/creditors.js`, an
untested API for creditors.  Integration tests specific to this
controller should be written following this PR.

It should be noted that our client uses some server API routes not
covered by integration tests.  These are:
 1. `/inventory/*` (used by patient invoice)
 2. `/creditors` (used by suppliers)

There is still lots of clean up work to do.  Here are some TODOs once
this PR lands:
 1. Begin using backticks (`\``) for multiline SQL queries on the
 server.  This will make the code much more readable!
 2. Ensure that the req.codes errors have been completed removed.  This
 is currently being worked on by @lomamech.
 3. Clean up confusion beteween `patients` and `patient_groups`.  There
 are some duplicate routes for listing routes.

Close #262.