feat: implement initial credit/debit sums
This commit implements the initial version of the debit/credit sum
indicators on the posting journal ui-grid.  There is still some work to
be done to ensure that the currency ids are available at startup.  A
subsequent commit should introduce auto-loading (and caching) currency
formats on application start for all available currencies.

Closes #290.