chore(voucher): update voucher schema
This commit updates the `voucher` table's schema to be more reminiscent of
the posting journal schema.  It adds in `entity_uuid`, and
`document_uuid` to the `voucher_item`, to allow greater flexibility in
attaching references to payments and other entities.  The integration
tests have been updated to reflect this change.

Closes #280.