feat(journal): add PostingJournalErrorHandler proc
This commit adds the PostingJournalErrorHandler stored procedure to the
procedures.sql file.  This ensures that all necessary SQL variables are
defined, and properly sends SQL SIGNAL errors if they are not,
terminating the transaction (and triggering a rollback).

The next step is to make these errors intelligible to the user via
translation keys.