feature(build): variable build times
This commit adds the ability to set the environmental variables
BUILD_TIMEOUT to control the time between building the repository and
initializing the server and the tests starting. This allows slow
machines to have a different build timeout from faster machines and
waste less time.  Defaults to 8 seconds as before.