test(invoice): formalize patient invoice API tests
This commit adds some simple tests for the patient invoices API and
improves the test documentation.

I've also created a helper function `helpers.mask(object, field)` to
temporarily remove a field from an object.  This makes it quite easy to
remove required properties from an object, resubmit it to the server,
and ensure that an error is returned.