test(invoice): add billing service + subsidy tests
This commit adds two more test for simple billing services (one billing
service) and simple subsidy tests (one subsidy).

In the process of testing these conditions, it was discovered that the
mysql `@` session variables do not get cleared at the end of the
transaction.  I've written a small utility `journal.cleanup()` to do
this process manually, but it is a hack.  In the future, we either need
to rewrite using only MySQL stored procedures or find a way to cycle
mysql connections to kill the session variables.