E2E For proposed debtor/groups page
This commit investigates running end to end tests with a notification
library setting a `$timeout` causing protractor to 'hang' or wait for
the `$timeout`.

* As E2E could not interact with the notification (given the
syncronization issue) the notification library was temporarily force
forked (copied) into the standar 'bh' library until either another
library is decided on or the current one is updated.

* Notifications moved into a temporary service as the model still existing
on controller (state) changed caused issues.

* Cleaned up temporary styles