Limit scope of ng-animate; update e2e tests
This commit addresses performance issues found during code review.
ng-animate was being applied to allow show/hide/if elements in BHIMA. As
these were not configured for animation there was a significant delay
between them being removed and dissapearing. The changes introduced here
limit the scope of ng-animate to only elements with the class
`ng-animate-enabled`.