Prototype filter $dirty value- nested components
This commit introduces a very basic method that allows the standard
`util.filterDirtyFormElements` to work with complex nested components
(usually represented as forms).

This method has been tested on the `bh-location-select` component;
noteworthy changes:
* Accept a `name` parameter (as per standard form elements) and use it
to name the form - this will become the attribute in the top level form.
* Whenever the components model value is set, expose this value to the
copmonent form through the value `$bhValue`
* Update util.filterDirtyFormElements to use `$bhValue` if
`angular.isDefined($modelValue)` is `false`

Misc.
* Move scope of group `ng-form` to ensure all elements are included