fix(login): ensure $locationChange respects login
This commit fixes a critical bug introduced throughout the `ui-router`
upgrade that allows users to access any page without logging in to the
application.  This has been patched so that $locationChange does not
even begin if the user session does not exist.

Finally, informative messages have been added to tell the user exactly
what is going on and what to do next when the session exists, doesn't
exist, or otherwise.