fix(login): prevent $state.go() getting to login
This commit ensures that $state.go() doesn't get to `/login` when the
user already has a session logged in.  Since $state-based and URL-based
routing are decoupled in bhima, we need to two checks to make sure the
URL is correctly displayed (and doesn't transition) and the $state is
correctly blocked and doesn't transition.

Fixes bugs reported in #222.