chore: remove fatal method from Notify
This commit removes the `fatal()` method from the Notify service - the
`fatal` messages should be reserved for errors that are out of reach of
the user and cannot be fixed without a sysadmin.  These messages now use
`danger` to render uniform errors.

It also removes welcome/goodbye messages from the SessionService.