Improve documentation with doclets.io (#383)
* feat(doclets): added .doclets.yml for auto-docs

This commit enables continuous documentation on doclets.io by adding in
the configuration file `.doclets.yml`.  Currently, only the server is
parsed and converted into documentation.

* docs(server): improve cash and patient docs

This commit improves the JSDoc syntax in cash and patient docs in order
to get better nesting of documentation.

* docs(server): add JSDoc comments to server/app.js

This commit adds JSDoc comments to the app.js file on the server and
restructures code to make the code clearer and produce better JSDoc
documentation.

* docs(db): add JSDoc comments to lib/db

This commit improves the database connector documentation by adding
JSDoc comments and a few more todos.

* docs: fix broken test documentation

This commit fixes broken api testing documentation by removing the JSDoc
syntax around `@fixme` and `@todo` comments.  It also removes the inline
comments from being JSDoc comments.

* docs(readme) update readme w/ doclets.io link

This commit updates the README with an API documentation link to the
doclets.io service for further browsing.

* fix(debtors): broken integration test on node <5

This commit patches a broken integration test that requires block-level
elements to be run in ES5 strict mode.  Whoops.

* docs(errors): improve server error documentation

This commit improves the server error documentation by updating the
documentation to JSDoc syntax, properly labeling functions, and
including examples for each function.

* chore: rm unused libraries

This commit removes deprecated and unused libraries `store` and
`validate`. Alternatives can be gotten from NPM if necessary.

 1. `validate` was an initial attempt at performing data validation,
 that should be replaced by the built-in validators from `lodash` or
 `mysql`.
 2. `store` was never used, only ported from the Dojo Store.  If it is
 ever needed, we can easily develop a new version or transition to
 WeakMaps in ECMAScript 6.

* docs(server): improve docs for routes/middleware

This commit starts a skeleton structure for improving the documentation
for routes and middleware.