chore(patients): restructure server architecture
This commit restructures the server architecture around the /patients
routes. The goal is to reduce duplication and the number of lines in the
bloated patients controller.

The new folder structure looks like this:
 - `medical/`
  - `patients/`
   - `index.js`
   - `groups.js`
   - etc...

Additionally, the `patient_groups` route is removed and replaced
with `/patients/groups`.  This does not change the controller at all -
patient groups are still created, updated, removed, and listed via the
patientGroups.js controller.

However, the URL structure is more pleasing to read:
 - `/patients/:uuid`
  Get the details of a single patient.
 - `/patients/:uuid/groups`
  Get the groups associated with the patient with uuid `:uuid`.
 - `/patients/groups`
  Get all patient groups in the database.

This commit adds the database structure and sample tests for patient
documents.  These will be implemented in a future commit.

Finally, winston logging has been improved for readability on uncaught
exceptions.