feature(uploads): implement file uploading
This commit implements file uploading using the multer middleware and
proves it using some example tests for the patient documents feature.
The entire server API has been implemented for this feature.

The uploader library function has been implemented to facilitate file
uploading.  It is a simple wrapper over multer that creates directories
using `mkdirp`, generates unique file names using `uuid.v4()`, and adds
a hardlink to the file to the req.file object.  For more information,
read the documentation in the uploader library file.

This adds a dependency on mkdirp.