chore(test): use browser.setLocation()
This commit changes all tests to use `helpers.navigate()`, which is an
alias for `browser.setLocation()` for navigation.  This has significant
performance advantages since it not longer requires app to download and
parse the page as `browser.get()` does.

Additionally, the incorrect JSDoc syntax `@params` has been replaced
with `@param` wherever it can be found.

Closes #380.