bug(session): ensure session is properly destroyed (#392)
This commit hooks up a listener to the HTTP 401 status code that
immediately destroys the user's session.  If necessary, we can embed a
logout directive in the HTTP response if this is too non-specific.  At
the moment, this fixes the bug that kills the application when the
server is restarted and the user still has their session.