feat(bhCurrencySelect): allow linking of cashboxes
This commit adds in an error message if the currency accounts have never
been set for a cashbox.  It provides a link back to the cashbox's edit
form to allow a user to easily update the cashbox's currency account
from the input directly.  This isn't displayed if the cashboxId is not
passed in.

It makes the code for opening the cashbox select modal a little more
clearer and improves the user interface slightly.  In particular, the
"do not see your cashbox?" prompt has been moved into the modal.

TODOs
 1. Update the cashboxes page with ui-router to allow deep linking from
 the cash page.  The links should be:
   a. From the bhCurrencySelect to the cashboxes/id/update page
   b. From the cashboxSelectModal to the cashboxes/create/new page.