test(voucher): create complex.page.js
This commit adds a page object representation for Complex Journal
Vouchers to facilitate testing.  In accomplishing this, the HTML of
complex journal vouchers is cleaned up to allow faster rendering and
better test selector targets.

Instead of individually selecting `[data-account-row="1"]`, each row is
given an index `[data-row="$index"]`, and used as the basis for all
row-level selects.

Two new features in FormUtilties:
 1. FU.typeahead() to support setting typeahead values
 2. FU.dropdown() to support setting dropdown menus

Closes #426.