feat(components): add ui-grid loading indicator
This commit adds a `bhGridLoadingIndicator`, a component that shows nice
messages on loading states, empty states, and error states.  The basic
usage looks like this:
```html
<div ui-grid>
  <bh-grid-loading-indicator
    loading-state="Ctrl.loading"
    empty-state="Ctrl.grid.data.length === 0"
    error-state="Ctrl.hasError">
  </bh-grid-loading-indicator>
</div>
```

The `hasError` state takes precedence over all other states - if this
value is `true`, it will not show any of the other loading indicators,
even if data has loaded or is loading.  The message is very generic -
one is expected to report the proper error using `Notify.danger()` or
`Notify.handleError()`.

For a real example, see the billing services module and patient
registry.